+++ spec/verifyCredentials.spec.js
@@ -1,10 +1,15 @@
 const { expect } = require('chai');
 const verify = require('../verifyCredentials');
 const {
-  table, unevenTable, unevenTable2, largeTable, duplicateTable,
+  table, unevenTable, unevenTable2, largeTable,
 } = require('./testData');
 
 describe('Verify tests', () => {
+  let eMessage;
+  beforeEach(() => {
+    eMessage = undefined;
+  });
+
   it('Parses successfully', () => {
     verify({ table });
   });
@@ -13,31 +18,26 @@ describe('Verify tests', () => {
     try {
       verify({ table: unevenTable });
     } catch (e) {
-      expect(e.message).to.be.equal('In row 1; Too few fields: expected 3 fields but parsed 2');
+      eMessage = e.message;
     }
+    expect(eMessage).to.be.equal('In row 1; Too few fields: expected 3 fields but parsed 2');
   });
 
   it('Recognizes an unrectangular table the other way', () => {
     try {
       verify({ table: unevenTable2 });
     } catch (e) {
-      expect(e.message).to.be.equal('In row 2; Too many fields: expected 3 fields but parsed 4');
+      eMessage = e.message;
     }
+    expect(eMessage).to.be.equal('In row 2; Too many fields: expected 3 fields but parsed 4');
   });
 
   it('Rejects CSVs that are too large', () => {
     try {
       verify({ table: largeTable });
     } catch (e) {
-      expect(e.message).to.be.equal('CSV is too large');
+      eMessage = e.message;
     }
-  });
-
-  it('Rejects CSVs with duplicates', () => {
-    try {
-      verify({ table: duplicateTable });
-    } catch (e) {
-      expect(e.message).to.be.equal('Duplicates exist in the English column');
-    }
+    expect(eMessage).to.be.equal('CSV is too large');
   });
 });